Создает модули в проекте
