+++ js/data.js
@@ -1,0 +1,13 @@
+const NUM_POOL = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11];
+
+const TYPE = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+
+const TIMES = ['12:00', '13:00', '14:00'];
+
+const FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+
+const PHOTOS = ['https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg', 'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg', 'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'];
+
+const SIMILAR_ADVERTISEMENTS = 10;
+
+export {NUM_POOL, TYPE, TIMES, FEATURES, PHOTOS, SIMILAR_ADVERTISEMENTS};
